requirements.txt
lxml==3.6.0
python-dateutil==2.5.3
urllib3==1.15.1
six==1.10.0
